♻️ refactor: stringifySize 유틸 함수 이동 및 기존 컴포넌트의 import 경로 변경
