+++ src/lib/utils/stringifySize.ts
@@ -1,4 +1,4 @@
-import { Size } from '../../types';
+import { Size } from '../types';
 
 export const stringifySize = (size: Size): Size => {
   if (typeof size === 'number') {
